Organize the source files in the OAuth package
