+++ packages/oauth/src/stores/index.ts
@@ -1,2 +1,3 @@
+export { InstallationStore } from './interface';
 export { default as MemoryInstallationStore } from './memory-store';
 export { default as FileInstallationStore } from './file-store';
