+++ packages/oauth/tsconfig.json
@@ -30,6 +30,7 @@
   ],
   "exclude": [
     "src/**/*.spec.js",
+    "src/**/*.spec.ts",
     "src/**/*.js"
   ],
   "jsdoc": {
